Do the same changes for midi
