+++ tracker/agb-midi-core/Cargo.toml
@@ -8,10 +8,6 @@ description = "Library for converting MIDI files for use with agb-tracker on the
 repository = "https://github.com/agbrs/agb"
 
 [dependencies]
-proc-macro-error = "1"
-proc-macro2 = "1"
-quote = "1"
-syn = "2"
 rustysynth = "1.3"
 midly = { version = "0.5", default-features = false, features = [
     "alloc",
