+++ tracker/agb-midi/Cargo.toml
@@ -14,3 +14,11 @@ proc-macro = true
 agb_midi_core = { version = "0.20.5", path = "../agb-midi-core" }
 proc-macro-error = "1"
 proc-macro2 = "1"
+
+quote = "1"
+syn = "2"
+rustysynth = "1.3"
+midly = { version = "0.5", default-features = false, features = [
+    "alloc",
+    "std",
+] }
